Remove console.log statements
