--- src/components/Data.js
@@ -1,34 +1,0 @@
-/* eslint-disable import/prefer-default-export */
-
-export const UserData = [
-  {
-    id: 1,
-    year: 2016,
-    userGain: 134118500,
-    userLost: 823,
-  },
-  {
-    id: 2,
-    year: 2017,
-    userGain: 78336300,
-    userLost: 345,
-  },
-  {
-    id: 3,
-    year: 2018,
-    userGain: 122207100,
-    userLost: 555,
-  },
-  {
-    id: 4,
-    year: 2019,
-    userGain: 115366700,
-    userLost: 4555,
-  },
-  {
-    id: 5,
-    year: 2020,
-    userGain: 142689800,
-    userLost: 234,
-  },
-];
