+++ src/components/Utils/API.js
@@ -1,4 +1,4 @@
 /* eslint-disable import/prefer-default-export */
-export const ApiKey = 'bbce26616140cb78fbac5081d8a07a35'; // this is a borrowed key'bbce26616140cb78fbac5081d8a07a35'
+export const ApiKey = 'fea98c7e52d82144069861c1eab1b036'; // this is a borrowed key'bbce26616140cb78fbac5081d8a07a35'
 export const API = 'https://financialmodelingprep.com/api/v3/';
 export const imgUrlGenerator = (code) => `https://financialmodelingprep.com/image-stock/${code}.png`;
